+++ Assets/Resources/FireMage/FavorOfFireStatus.prefab.meta
