+++ Assets/Scripts/Status/BASE_STATUS/IBuff.cs
@@ -1,1 +1,5 @@
-﻿public interface IBuff : IStatus { }
+﻿/** IBuff, public interface
+ * Grouping interface that concerns Buffs (Empowerments)
+ * Should be used with StatusBase only.
+ **/
+public interface IBuff : IStatus { }
