+++ Assets/Scripts/Status/BASE_STATUS/IDebuff.cs
@@ -1,1 +1,5 @@
-﻿public interface IDebuff : IStatus { }
+﻿/** IDebuff, public interface
+ * Grouping interface that concerns Debuffs (Damages or curses)
+ * Should be used with StatusBase only.
+ **/
+public interface IDebuff  {}
