+++ Assets/Scripts/Status/BASE_STATUS/IDebuff.cs.meta
@@ -1,6 +1,6 @@
 fileFormatVersion: 2
-guid: f6605458c61acd54c89402b59232f4bc
-timeCreated: 1511897238
+guid: 38483362ed23ad84fbb80b61ab673731
+timeCreated: 1512328731
 licenseType: Free
 MonoImporter:
   externalObjects: {}
