+++ Assets/StreamingAssets/StatusData.json
@@ -1,6 +1,6 @@
 [
 	{
-		"ScriptName":"FireBlessingStatus",
+		"ScriptName":"FavorOfFireStatus",
 		"Name":"Favor of Fire",
 		"Element":"Fire",
 		"Duration":5,
